+++ src/aioway/__init__.py
@@ -5,7 +5,6 @@
 from . import (
     _api,
     _costs,
-    _iters,
     _sess,
     _thunks,
     _utils,
