+++ src/aioway/io/sinks.py
@@ -6,7 +6,7 @@ import abc
 import dataclasses as dcls
 import typing
 
-from aioway._iters import Iter
+from aioway.relalg import Iter
 from aioway.spaces import AnySpace, Space
 
 __all__ = ["Sink", "sink_dcls"]
