+++ src/aioway/relalg/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
+from ._iters import *
 from .funcs import *
 from .joins import *
 from .loaders import *
