+++ src/aioway/relalg/_iters/iters.py
