+++ src/aioway/relalg/_iters/nodes.py
