+++ src/aioway/relalg/views.py
@@ -9,8 +9,7 @@ from collections import abc as cabc
 import tensordict as td
 import torch
 
-from aioway._iters import TdictIter, TensorIter, node_dcls
-
+from ._iters import TdictIter, TensorIter, node_dcls
 from .maps import MapIter
 
 __all__ = ["ColumnViewIter", "ProjectIter"]
