+++ tests/test_emitters.py
@@ -5,12 +5,11 @@ import typing
 import pytest
 from torch import nn
 
-from aioway._iters import TensorIter
 from aioway.attrs import Attr, Shape
 from aioway.compound import BuiltModule
 from aioway.emits import MlpEmitter, emit, emit_one, linear_shape
 from aioway.io import TensorListIter, TensorStream
-from aioway.relalg import LoaderOpt
+from aioway.relalg import LoaderOpt, TensorIter
 from aioway.spaces import AttrSpace, ShapeSpace
