+++ tests/test_indices.py
@@ -3,10 +3,10 @@
 import pytest
 import torch
 
-from aioway._iters import TensorIter
 from aioway.attrs import DType
 from aioway.indices import AnnIndexIter, AnnIndexTrainerIter, FaissIndex
 from aioway.io import TensorListIter
+from aioway.relalg import TensorIter
 
 
 @pytest.fixture
